Update FQuiz1_Recovery.docx.md
I fixed the header. By the way, this quiz is all about the case study of  Cybersecurity Resilience: TechGuard Solutions' Recovery Disk Strategy in Action. I learned a lot here because I did get some ideas on what to do when one of our computers will be infected with virus and such.